successfully test code from last fall
We haven't been able to duplicate the results from last fall, in terms
of performance. This commit contains code which does so while also
respecting boundaries between vocabulary and test categories, and using
a single set of models. Future commits will be testing ways to recreate
these models. Perhaps the important thing here is that we used the
"organicC3" and "inorganicC3" patchsets cached as "organicC3OldSchool"
and "inorganicC3OldSchool".