simulation with newly created, non-mined models
This is the third in a series of simulations. The first showed that we
get stellar results when the training and testing categories overlap,
which isn't a useful result. The second showed that we get good results
when those sets are disjoint and when we used the models generated last
summer. This third simulation shows that creating new C3 models from
scratch without hard negative mining, we have approximately similar
performance. The fourth simulation will test new models created with
hard negative mining.