fifth simulation, all new random decisions
This fifth simulation builds the conclusions of the past four, that we
have good results without mining using this codebase, to generate
results based on entirely new random decisions. Everything from the
decision of which categories to include is completely new. But, the
effect we're looking for disappears. So, I have some work to do in
figuring out where that result has gone.